cxl: add support to build and install cxl_test
The cxl_test module is an out-of-tree but-in-tree driver part of
the Linux kernel which uses __weak attributes to do some mocking
of different sorts and then allows the ndctl meson unit test to
go to town with it.

In order to suport the meson unit test stuff from cxl we must first
build/install the cxl_test driver. Add support for that and allows
helper targets which lets you build/install *just* the test_cxl
driver alone too in case you are doing some development with that.

  * make cxl-linux
  * make linux-cxl

These are the same thing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>