cxl: use INSTALL_MOD_DIR=updates when installing the mock modules
The cxl_test driver replaces existing CXL modules by using mocking code,
which essentially needs to load alternative modules instead of the
production ones.

Sadly without an upstream patch not yet merged if you *do* load the
production cxl modules *and* the cxl_test driver you'll crash with
a null pointer dereference. A fix to ensure this does not happen is
in being discussed now, however to ensure we properly load the mock
modules which are needed we just need to use the INSTALL_MOD_DIR=updates
paremeter for modules_install.

That will ensure those symbols are used over the produciton ones.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>