Kconfig: enable sensible defaults for using git mirrors
Heavy users of kdevops will want to use the local mirror setup.
This essentially mirrors large git trees used such as linux-next,
linux, linus-stable and qemu locally *once* so your more than one
delployment of kdevops can insantiate guests *fast*.

A default of just not enabling this does not make sense and means
we have to go and always enable this, and this can be pretty
error prone. Instead, just check to see if /mirror/ is present
and its not empty make the default to "y" to use the mirror.

Installing the mirror systemd service and timer files requires
a slightly different heuristic: enable installing if you've
selected this is your first kdevops run *and* have the /mirror/
directory present. Users without /mirror/ will simply not get
the mirror stuff installed. Folks who have it likely want it
installed.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>