vagrant/libvirt: add smart storage pool heuristic
A power usage of kdevops means git cloning it on many directories,
one per goal. So for instance I have:

/dev/nvme0n1 with xfs   on /data1/
/dev/nvme1n1 with btrfs on /data2/
/dev/nvme2n1 with ext4  on /data3/

And then I test xfs on /data1/ on linux-next so I have:

/data1/next/xfs/kdevops/

But likewise I also test XFS using btrfs on the host, so I have also:

/data2/next/xfs/kdevops/

And likewise I test xfs on ext4 on the host too... so I have:

/data3/next/xfs/kdevops/

But I also test CXL so I have:

/data1/cxl/kdevops/

And so on. To make most efficient use of the nvme drives I want my
guest images to be local to the same nvme drive so I create a data1,
data2 and data3 storage pool, and I allow kdevops to create these
storage pool paths for me. Once created though I want kdevops to infer
the pool name and path for me as I want to be lazy and I want sensible
defaults.

And so, this adds precisely that heuristic of looking for existing
pools and if they exist looking to see if the base directory I am
in has a directory which a storage pool mentions. If so then it uses
that storage pool name and path.

I only enable this for DISTRO_DEBIAN, and while this should work for
other distros, folks can easily modify their own distro default once
they verify it works there too. I add these guards to ensure I don't
break any other distro workflow by default.

All distros should eventually migrate to this new heuristic.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>