bootlinux: move 9pfs options before selecting git mirror
You want to be able to enable / disable 9pfs *before* you select
the 9pfs path for the mirror. Otherwise it will not make senese
the path will be a local to the host for a git clone. The path
to the git clone is local *if* you are enablign 9pfs but also
also have a local mirror of linux.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>