blktests: fix debug redirect for caching actual check command
To enable debugging when kdevops fstests or blktests is used,
since both use their own oscheck.sh, prior to calling their
respective test suite check script, it outputs to temporary
file the exact command that will be executed next on the
target node. The file is /tmp/run-cmd.txt

So you can always just cat the file after you run:

make fstests-baseline
make blktests-baseline

To help you debug a run in case of failure. The re-direct to
the output file was not set so we were sending the command out
to stdout. So this does not fix any run time issue, it just
fixes the debug output file actuallly existing on blktests.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>