Add simple minimal binary / program version check
The biggest "bug" I've been reported is about users
having odd issues with ancient version of ansible or
other tools.

Just make the requirements explicit so we can bail out
early instead of dealing with really obscure bug reports
for things that really just point the user to upgrade,
but they cannot see that, nor it is obvious to them.

This just ensures sanity for our minimum dependencies.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>