linux-mirror: fix mirror user experience
Right now the mirroring works but after upstream gets updated
our mirroring does indeed fetch the updates, but we don't move
the head so a clone without specifying a branch gets us the git
tree when it was first cloned.

The missing part was:

git fetch origin +refs/heads/*:refs/heads/*

Using git clone --bare works but it does not fetch the git notes
and some other attributes. We want it all so use --mirror instead.
When using "git clone --mirror" typically the recommended next step
to update a git mirror is just:

git update --prune

However, the documentation doesn't talk about fetching all the tags
and linux-next has tons of tags. By default only the tags for the latest
are carried over. It *may* be that "git update --prune" brings in all
the git tags, but I am not too sure of that yet so we add a secondary
call to ensure we have all the tags.

The last command we issue is:

git fetch origin +refs/heads/*:refs/heads/*

To ensure we move our updates our refs with what presente remotely.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>