nfsd: move nfsd startup to bringup phase
Currently it's started by "make pynfs", but future workflows will
also require it, and it's nice to be able to do by-hand testing with the
dev box. Enable it as part of "make bringup" instead.

This also ensures that the nfs-server service is enabled too so that
you can iterate using "make linux && make pynfs-baseline".

Signed-off-by: Jeff Layton <jlayton@kernel.org>