fstests: prefer enabling fstests watchdog and kill option
The    HAVE_DISTRO_PREFERS_FSTESTS_WATCHDOG and the respective kdevops
option HAVE_DISTRO_PREFERS_FSTESTS_WATCHDOG_KILL have been tested for
a while with stable / linux-next. Enable this by default so to enable
kernel-ci to make use of it.

Without this you can end up with a hung system and the systemd watchdog
will not pick up on it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>