README.md: annotate bug known when bringing up linux prior to fstests
Document a known bug which requires us to `make fstests` prior to
`make linux`. The issue you'd run into is not being able to mkfs
the nvme drive for the sparsefiles, and the issue is caused by
having done a reboot after having installed linux and used an
nvme drive for the /data parition and this same partition now
being /dev/nvme1n1. We need a way to associate data_device and
sparsefiles_device to something less non-deterministic like
maybe a serial number.

Note that we already associate with vagrant the data partition
with information of its purpose:

https://github.com/mcgrof/kdevops/blob/master/playbooks/roles/gen_nodes/templates/kdevops_nodes_split_start.j2.yaml#L57

The scratch device:

https://github.com/mcgrof/kdevops/blob/master/playbooks/roles/gen_nodes/templates/kdevops_nodes_split_start.j2.yaml#L59

*and* we actually already extract the purpose, ie, this "data" or
"scratch" when we're creating the nvme drives for qemu:

https://github.com/mcgrof/kdevops/blob/master/playbooks/roles/gen_nodes/templates/Vagrantfile.j2#L337

So we *could* already even associate this information to the serial
number easily of the device, it would just mean that we'd have
to extend the create_partition task to support 'disk_setup_serial_number'
as an alternative to disk_setup_device:

https://github.com/mcgrof/kdevops/blob/master/playbooks/roles/create_partition/tasks/main.yml
https://github.com/mcgrof/kdevops/blob/master/playbooks/roles/create_partition/defaults/main.yml

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>