workflows: document we need 8 GiB RAM for selftests
Document we need at least 8 GiB of RAM for selftests.
Note that I had tried something like this:

  --- a/kconfigs/workflows/Kconfig
  +++ b/kconfigs/workflows/Kconfig
  @@ -143,6 +143,7 @@ config KDEVOPS_WORKFLOW_DEDICATE_PYNFS
   config KDEVOPS_WORKFLOW_DEDICATE_SELFTESTS
          bool "Linux kernel selftests"
          select KDEVOPS_WORKFLOW_ENABLE_SELFTESTS
  +       select VAGRANT_MEM_8G if VAGRANT
          help
            This will dedicate your configuration only to Linux kernel
            selftests work.

But this *does* not work on kconfig today. kconfig has a real upstream
bug where this does not update to reflect the selection so there lies
a semantic gap. Then, also we'd just cover the requirement for vagrant
and not cloud.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>