cxl: Add second cxl topology
Add another root port to a single host bridge. A single persistent
memdev is connected to one of the root ports. This extends the
first demo topology. Note that the first demo topology represents
a passthrough decoder in kernel speak. The topology added in this commit
is no longer passthrough because the hostbridge has two root ports.

Signed-off-by: Adam Manzanares <a.manzanares@samsung.com>