scripts/libvirt_pool.sh: fix null re-direction orderign
Last fix... I hope is to redirect properly otherwise we get the failure
check on the output. Also while at it update the documetnation to
reflect how the heuristics are hopeful deafults for lazy developers
we wish to strive to, but for now we only enable it on debian. Even
if you are using debian the heuristics provide safe results if you
can't use sudo without a password so in that case the heuristics kconfig
option is still enabled, it is just the normal defaults will be provided.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>