pynfs: add infrastructure for kernel-ci loops
Wire up the pynfs workflow into the kernel-ci. Add a run_loop.sh
script and the necessary symlinks. For pynfs, we just have a single
set of baseline results in json format that is committed to the repo.

This also adds a python script for interpreting test results vs. the
current baseline. For now, the baseline is generated from the latest
bleeding-edge nfsd code. We'll plan to update it as the "gold standard"
changes.

Signed-off-by: Jeff Layton <jlayton@kernel.org>