kdevops: change the /data and /media definitions to use by-id paths
If you do PCI passthrough of other NVME devices into the qemu guests
you're going to have a bad time with the default settings.  We have the
/dev/nmve paths hardcoded based on what a clean vm looks like without
any PCI passthrough.  Fix this by using the serial option to have the
kdevops prefix so we can easily identify the devices we create for the
VM's.  Then use the /dev/disk/by-id/ paths in our configuration to make
sure we consistently get the correct device when setting up the /data
and /media partitions.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>