fstests: unset RUN_SECTION before running xfstests
If we're using the special "all" option that lets xfstests just run
we'll fail to run anything because it's set in RUN_SECTION, which is
used by xfstests to determin which section to run.  Unset it before we
call ./check, if we had it set we'll have added the proper option to the
command string.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>