oscheck: handle "all" fstests section better
Right now we just sort of by-pass the normal checks if we do "all" for
the fs section.  This causes weird failures when somebody is stupid like
me and messes up their configs.  Handle this by pulling the all of the
sections and validating them to make sure we fail at a sane place.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>