gen_nodes: add virtio and explicit IDE support
This completes the abstraction of the extra storage drive when
vagrant with libvirt is used. The same PCI bus hierarchy is followed
so that we can keep the extra drives in its own port and we don't
conflict with generic devices.

ZNS obviously cannot be enabled when virtio is used as we need NVMe
for that so we make this dependency strong with kconfig.

IDE was used to support older SLE releases which don't have virtio,
those older releases need to be adjusted so that they *select* the
IDE libvirt option.

I've tested this with a 'make fstests' target for both btrfs and xfs
with libvirt nvme and virtio. I haven't bothered testing ide yet.

The reason we are adding virtio support is to help compare and contrast
when we find issues with the qemu hw/nvme/* code, or we at least suspect
we do. The current lingering issue we are seeing is nvme timeouts. Using
qemu 7.2 should fix a lot of them, but we are still seeing some
timeouts. Some other enhancements are coming through for hw/nvme which
probably will do away with some of these last annoying issues.

Anyway, this should enable folks who really cannot use nvme for
whatever reason too. We make nvme the default and we expect this
to remain this way forever.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>