gen_nodes: replace nvme_disk_path with an ansible extra_disk_path
In order to abstract NVMe out of the extra disks provide a variable
abstraction for the extra disk path using a generic name. There is
no need to keep this in the yaml input file, the less we have there
the better so just favor the ansible defaults file instead.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>