vagrant: only use an ansible variable for the extra disk postfix
Right now we have two yaml variables we use for the nvme_disk_format,
and we modify it if its raw for libvirt. For virtualbox we always use
"vdi" even though two other formats are actually available / suppported
by virtualbox.

Rename the variable to be agnostic to the backend driver, and use only
one ansible variable for it all, and add Kconfig entries for the other
possible format for virtualbox.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>