pynfs: run the tests as root
Some things (like creating devices) can only be done as root, and most
of the known failures are due to this. Add the right options to
impersonate root access, and add "no_root_squash" to the default
parameters.

Also, add a new set of baseline test results, and change the "current"
links to point to them. With this, there is only one known failure under
NFSv4.0; the LOCK24 test.

I have a pynfs patch in flight to fix that, so we should be able to
update the results again once that's in.

Signed-off-by: Jeff Layton <jlayton@kernel.org>