9p: disable cache to fix client synchronization
kdevops embraced using a cache=loose through the qemu wiki [0]. However
this yields an unexpected terrible user experience:

Even if you do a full rebuild of Linux on the host after modifying
the top level Makefile the client *still* does not get any single
updates toe the top level Makefile.

This is insanity and it requires a full reboot of the guest in order to
actually see full changes of the kernel on the host. This is insanity
for kernel development.

Just replace cache=loose with cache=none then. This issue is currently
being discussed on the v9fs mailing list [0]. Next I'll try to go fix
this on the qemu wiki as the current recommendation is ill advised.

[0] https://lkml.kernel.org/r/ZCHU6k56nF5849xj@bombadil.infradead.org
[1] https://wiki.qemu.org/Documentation/9psetup

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>