vagrant: add new kdevops debian-xfs-20230427
It turns out that if you want to avoid using buffer head you want
to not be using filesystems which rely on buffer heads. Most vagrant
images use ext4 by default. We need an alternative so this provides
one as an example and for R&D development for large-block efforts
on the Linux kernel.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>