kdevops: adds support for pure iomap
struct buffer_head and its respective fs/buffer.c has historical
baggage which only recently can we seriously start considering
abandoning it with a proper replacement. Support for this is
only available in development branches for code not upstream yet.
Likewise, in order to use avoid using buffer heads you must also
be on a system which does not require filesystems or block drivers
which require support for buffer heads. Today this means images or
cloud instances which use an XFS or btrfs root filesystem.

Add support for this and add kconfig symbols which development
branches can select for support for indicating they have these patches,
and allow vagrant images and cloud instances to also indicate they are
safe to boot in these environments.

Most vagrant images for the latest distros use at least an vfat
partition /boot/efi. So sadly pure-iomap cannot be used on those
either, even if they did use xfs for the root OS image. We provide
a kdevops image based on debian testing which shows removing vfat
works on boot for testing and R&D. On the cloud I see the same issues.
The only cloud instance to support only and *only* XFS at boot is
AWS 2003 instances.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>