fstests: remove 6.3.0-next-20230426 expunges
The tests were run on a qemu 4kn drive but the loopback
drivers created were using 512 logical block size. We need
to fix losetup to not allow for that. It must match the
logical block size of the backing drive.

So all the issues are likely test issues. Just nuke it all.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>