create_partition: wipe fs on stale drives
Ensure to wipe any filesystem on a drive where there is no mounted
target for the drive and we are about to mkfs on it. This would have
captured old stale data on drives where say an instance is brought up
but using a copied or old stale extra drive. This shouldn't happen
anymore due to the recent commits but this would let us know if
such situation happens early on.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>