create_partition: fix label check for network filesystems
The label check for partitions is block specific, we however
now suport 9p for the data partition and so must check for
all possible devices, not just block devices.

Without this you can end up creating the /data/ partition for
say 'make fstests' but then fail when trying to run 'make linux'.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>