vagrant-libvirt: add support for s3 and s4
To see if a virtual platform has support for s3 or s4 you can use:

$ sudo irsh qemu-monitor-command vagrant_kmod 'query-current-machine'
{"return":{"wakeup-suspend-support":true},"id":"libvirt-415"}

However, that's not enough. You could sprinkle this on your xml after
the <os> section:

  </os>
    <pm>
    <suspend-to-mem enabled='yes'/>
    <suspend-to-disk enabled='yes'/>
  </pm>

But that still doesn't tell me what to do when bringin up qemu.
This happens to be machine specific and I can only infer that
for q35 it's disabled by default... And so to enable you gotta
force enable it.

So to enable we have to twiddle a disable bit for each machine, for
q35 we add:

  -global ICH9-LPC.disable_s3=0 -global ICH9-LPC.disable_s4=0

To test suspend on guests just use:

$ sudo systemctl suspend

And to wake the guest up on the host:

$ sudo virsh dompmwakeup vagrant_kmod

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>