bootlinux: Do not reboot after linux-install
Do not reboot after linux-install. This target builds and deploys the
kernel into the VM but does not configure the GRUB entry to use it,
so it requires the user to manually select the new GRUB with `virsh
console <num>`.

A user can simply run `make linux-install; make linux-grub-setup;
make linux-reboot` to automate this. Or `make linux-install; make
linux-reboot` to achieve the previous behaviour.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>