Fix bootlinux target with heterogeneous host/guest OS
Currently, the bootlinux role tries to install dependencies on the host
using the method that would be appropriate for the guest. Upshot: you
couldn't call make linux with heterogeneous host and guest OSs.

Add a completely new playbook that we can run with a local connection on
the host, with a separate command.

Signed-off-by: Jeff Layton <jlayton@kernel.org>