Add updated linux-next config
This also turns off CONFIG_MODULE_SIG as well, and all of the ripple
effects from that:

The sign_file program that make modules_install uses to sign the modules
is built as a dynamically linked program during the build phase. This
means that if you try to run it on a different os_family, it can fail
because of the linker differences.

Given that we're interested in testing kernel changes here, module
signing doesn't add anything of value and can get in the way of
debugging at times (if you want to plug in a replacement module, for
instance). Let's turn it off in the latest linux-next bootlinux
template.

Signed-off-by: Jeff Layton <jlayton@kernel.org>