Fix disablement of HV tuning
If you're running in a cloud provider, you don't necessarily want to
tune the local hypervisor. If hypervisor_tunning_enabled is false, then
we currently don't attempt to stat the files, so the appropriate dicts
for ksm_enable_file and zswap_enable_file never get created.

Always stat those files, so that the dicts get created whether they are
used or not.

Signed-off-by: Jeff Layton <jlayton@kernel.org>