terraform: fix ssh config update rationale
Updating your ssh configuration when using terraform used to work back
in the day for all providers. However right now this is known to not
work on both AWS and Azure. On OCI it certainly works for some reason.

Fix all this by making it very clear what variable we want to use for this
and use a terroform bool. Note that terraform bools are lowercase "true"
or "false, and so they differ from Ansible True and False. So to use
ansible variables wer just {{ foo | lower }}.

Terraform folks don't plan to change terraform's "bool" type [0], and so
since it is fixated, just play along.

This:

  * simplfies the variable setting of update_ssh_config_enable on the module
    "ssh_config_update_host_entries" to be directly set by
    var.ssh_config_update. All of the prior commits were done to clarify
    the odd stacked set of variables used here before which were just
    redundant.

  * stick to using Ansible semantics when using ansible, and make it
    clear we disable these variables by default and theen only enable
    them if enabled on kconfig.

  * the terraform var.ssh_config_update is set to {{ foo | lower }}

  * fixes a typo on "devops_terraform_ssh_config_update_backup="

  * removes 3 extra unused duplicate variable names,
    kdevops_terraform_ssh_update and friends.

[0] https://github.com/hashicorp/terraform/issues/24354

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>