terraform: fix strict & backup settings
The less terraform does the better, and parsing is one of the major
small issues I see with where bugs can easily lurk. By reducing the
level of idirections we fix forcing strict ssh settings and the backup.

The strict settings are borrowed in style from terraform's default
ssh settings.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>