nfsd: create a dedicated host for it when the server is enabled
After working with kdevops a bit, I think the way I've implemented the
NFS server is wrong. Instead of making it a feature of any host, bring
up a dedicated node as an NFS server. Add a new "nfsd" group in the
ansible hosts file and only do the nfsd setup operations against that
host.

Signed-off-by: Jeff Layton <jlayton@kernel.org>