LBS: force VAGRANT_KDEVOPS_DEBIAN_TESTING64_XFS_20230427
LBS devices require pure iomap for now, and no vagrant image for
standard distributions provide the requirements for that, to use
only XFS or btrfs as the root image. So force our own vagrant image
for now if using vagrant. If you are on the cloud you can experiment
with large-block-20230525 by using Amazon Linux 2023 on AWS too.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>