cxl: make jic23 the default qemu tree
qemu builds with the upstream v8.0.0-rc1 yields an environment where
booting even a stable kernel works but upgrading after to a kernel
with commit a49aa8141b65 ("cxl/mem: Wire up event interrupts") you
will end up with something like:

"cxl_pci 0000:0d:00.0: Failed to get interrupt for event Info log"

This is a known issue but [0] but it would seem this is a requirement
for CXL anyway.. so users have no option but to get a stick and are
forced to upgrade CXL.

This is stupid but oh well. So just follow Jonathan Cameron as the
true north star for CXL and use the latest branch there.

[0] https://lore.kernel.org/all/646793cc665bf_1231462943c@iweiny-mobl.notmuch/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>