vagrant: don't create a :type => user interface for qemu:///session
When I create a guest in vagrant under qemu:///session, I currently get
two interfaces. One is bridged to the local default libvirt network,
and the second is a :type => "user" network. This second network seems
to have no purpose all of the "real" work is done using the default
bridged interface.

I think the FAQ indicated in the comment is out of date vs. modern
distros, at least with QEMU/KVM. On my F38 box, /etc/qemu/bridge.conf
contains this:

    allow virbr0

...which allows any guest to connect to the default virtual bridge.

Remove this section of the Vagrantfile, which doesn't seem to add
anything useful.

Signed-off-by: Jeff Layton <jlayton@kernel.org>