terraform: add option for private network
We eventually want to enable inter-host communication for some tests.
This is not usually a problem with Vagrant since those environments
usually operate behind a firewall, but with hosts in public clouds, we
can sometimes end up with a public IP address.

Most cloud providers have a way to set up a private network, and some of
the terraform configs already set one up by default. It's not clear to
me whether that can carry extra costs however, so add a Kconfig option
for this, and make nfsd depend on it when terraform is enabled.

For now, only openstack is wired up for this. We'll add support for
other providers in later patches.

Signed-off-by: Jeff Layton <jlayton@kernel.org>