gendisk: capture the environment variable if it is set
kdevops was giving out the following error:
TEST_DEV environment variable is required to be set when the -m argument
is used", "TEST_DEV: "] even though the TEST_DEV env var was set.

Change gendisk.sh to capture the env var if it is set instead of
overriding it all the time. This fixes the issue and creates and formats
the loopback devices properly.