xfs: add kconfig options for LBS with 4k sector size
Enabling true LBS devices is nice, ie, where their logical block size is
greater than 4k. But we want first a baseline to see what explodes
first, so add profiles which allow us to enable large block sizes with
a sectorsize of 4k. This means to test these you'll want to customize
kdevops so that the qemu drives come up with 4k instead of the default
of 512 bytes.

To help make things clearer, add an option which clarifies what it
means to have a real LBS device, it's one where it is capable of
having a physical block size of at least the large block size.

Note that some storage drives are capable of a large physical block
size even if the LBA format is still 4k.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>