fstests: place fstests install on the /data/ partition
When testing XFS with LBS it is *immediately* clear failures will
produce much larger XFS metadata dumps, we're talking about ~ 1 GiB
for a test fail. The default vagrant guest image uses about 20 GiB
for the / partition, and that means after about 20 or so failures
we'll run out of disk space on on / because /var/lib/xfstests/results
is being filled fast with failure logs / XFS metadumps / etc. This
makes testing with LBS not possible today.

To fix this move the install target to use the /data partition which
has much more space, 100 GiB by default. If we need more than that
clearly the testing results are so bad that obviously the filesystem
code needs more love before you proceed to test more.

So we install fstests onto /data/fstests-install/xfstests and use
a symlink to point /var/lib/xfstests to this directory. We remove
all of them prior to installing fstests.

To use this new directory we must configure fstests using --prefix,
and so we must first build the configure target for fstests first now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>