fstests: add support for using real NVMe drives
kdevops supports using loop drives with sparse files well for
a long time now. But automatic and intelligent schemes for
when you have real NVMe drives have been lacking. For ZNS we
ended up assuming you can just punch in the menu the actual
drives you could use. This is boring.

If you want to test with real NVMe drives you probably don't
want to run tests against a real host but instead use PCIe
passthrough and so you can then blow up the guest kernel and
not worry about crashing your host.

One of the problems with using real NVMe drives or even with
PCIe passthrough is upon reboot the drives may re-order. To fix
this we must use the /dev/disk/by-id/nvme.eui.* symlinks. To
use these we must extend fstests to support symlinks and then
canonicalize them. I have a patch to do just that and will post
soon.

This work adds two new schemes to let you use existing NVMe
drives on a real system or target guests with PCIe passthrough
so that can either use either:

a) the first 7 NVMe drives found with a proper eui
b) the first   NVMe drive  found and we create 1 extended
   and 12 logical partitions so you can test fstests on

The ansible tasks for the partition magic use parted and we
default for now to 60 GiB size partitions. This updates the
documentation for the sparse file generation also while at it
so to explain the existing ordering scheme a bit better, where
it comes from, and so to also pave the way to extend this and
the new real NVMe drive schemes a bit easily later.

I've tested this with the partition logic creation. The only
issue I see is sometimes you wipefs -a /dev/nvme0n1 and then
the creating the extended partition fails, wipe'ing again
and re-running 'make fstests-config' fixes that, but this
is an odd bug, even though parted says its a parted bug.

I hide this magic under a new FSTESTS_TESTDEV_UNDER_DEVELOPMENT
for now as this is still work being worked on, and only if you
are happy to find bugs should you use this stuff. You will obviously
also need some non upstream patch for fstests which I'll post
soon which lets you use symlinks.

I've tried to be careful as I can about ensuring this doesn't
break existing setups.

If you want to use this on another filesystem you want to
use RECREATE_TEST_DEV=true just as we do in XFS, but also
we may just want to see if we enable this in the future on
all filesystems even for sparsefile generations and then
remove the gendisks.sh and oscheck.sh ability to create
the first TEST_DEV partition for you. We'll need folks
running tons of tests with kdevops for feedback to ensure
this doesn't break their setup and confirmation that this
simplification is this is good.

The other odd thing we should later review is all the environment
variables we stuff under the task which runs the final oscheck.sh.
Although that was added for some odd distro sudo environment,
do we still need it? oscheck.sh has been extended considerably to
scrape the config/*.config for variables now before running, so
maybe this is not needed anymore? We should probably also just
generalize how we do that and upstream onto check.sh that logic
so everyone can benefit from trimmed down filesystem configurations
with one generic shared section.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>