fstests: provide a fallback mechanism for NVMe drive hunt
As per the NVMe base specification 2.0c under section 3.2.1.6 NSID and
Namespace Usage, devices which support namespaces, ANA reporting or
NVM sets need to pick one of NGUID, EUI64 or UUID to implement and
suport. So EUIs are not required and not all drives support all these
features. So some drives may exist which don't use EUIs. To help with
this this allows you to use a fallback of the Linux kernel symlink
fallback of:

/dev/disk/by-id/nvme-$(model)-$(serial)

This mechanism will by default skip qemu NVMe drives. We also
enable it by default if you enable NVMe drive scanning as this is
the only way to use most drives.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>