fstests: add support to use check.time files as expunges
One of the hardest things to do is to work on an initial baseline
for a filesystem. This requires thankless work to find all issues
first, write down each failure / log them. This allows you to them
later test a new delta patch to see if you've introduced new
regressions.

One of the current issues with this effort on kdevops is we typically
run tests again, by adding to the expunge list for the kernel the
known failures. One of the issues with this is we're re-testing
all existing tests that ran before. We could modify upstream
fstest to "start" testing after a certain test number, but that
implies you are testing without a random order option. That's
fine but there is another shortcut we can use for now, we could
simply use the existing prior check.time files for expunges for
tests, so that new runs will skip tests which already ran at least
once if they passed.

Add initial support for this using the environment variable
INITIAL_BASELINE, and so you could use:

make fstests-baseline INITIAL_BASELINE=1

We'll experiment with this a bit and later see if we do more
modifications to it. One of the issues with this is we have
to monitor to ensure old check.time files are kept.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>