bootlinux: move kconfig files to workflows/linux/
Although the ansible role is named bootlinux, the workflow is
clearly Linux so just move the Kconfig files there. The name
"bootlinux" was used as that is what we try to do, configure,
build and boot into Linux.

We could probably later just rename the ansible role if we want
to just linux.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>