fstests: add default nvme_euis
Add a default empty list for nvme_euis in case there is none anyway.
Then only remove the main drive partitions if we set out to create
partitions and we don't see what we expect, the 13 partitions.

Also adjust the creation of logical partitions to indicate we went
to create 12, we only create 12 logical partitions, but also create
a first extended partition.

Fixes: b5869aa515695 ("fstests: add support for using real NVMe drives")
Fixes: 1e7c8f380d6ed ("fstests: provide a fallback mechanism for NVMe drive hunt")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>