CXL: Add switch topology
Add support for a cxl switch in the topology. One upstream
port connected a host bridge with a single root port. Two
downstreams ports with persistent memory device attached
to one of the downstream ports.

Simplified the conditionals in Vargrantfile.j2 because much
of the template is shared between all topologies.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Adam Manzanares <a.manzanares@samsung.com>