xfs: add initial baseline for next-20230720 on 4k blocksize
Although this test is for the 4k reflink section we just provide it
as the all.txt expunge so all sections uses it for now.

This tests linux-next tag next-20230720, ie, 6.5.0-rc2-next-20230720
on XFS with 4k block size using qemu sparsefiles on an nvme drive
configured with physical and logical block size to 4k, so 4kn drives:

CONFIG_QEMU_NVME_ZONE_PHYSICAL_BLOCK_SIZE=4096
CONFIG_QEMU_NVME_ZONE_LOGICAL_BLOCK_SIZE=4096

The confidence in this baseline is just one loop due to time, but
we can increase it now that we have something to work with.

At least two crashes should be removed once we move to
next-20230724 or next-20230725 as an x86 regression is fixed there.

While at it, provide some guidance as to how to easily view
errors when we commit the results archive in xz tarball format
without decompressing the full file to disk, ie, only get the
output for a test *.out.bad file out to stdout. Later on we can
make tools to make this easier. A future dashboard could then
view the full output with this command too.

For example to view the actual generic/471.out.bad output:

tar -xOJf \
 workflows/fstests/results/mcgrof/libvirt-qemu/20230725/6.5.0-rc2-next-20230720.xz \
 6.5.0-rc2-next-20230720/xfs_reflink_4k/generic/471.out.bad

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>