xfs: add initial full baseline for next-20230725
The baseline for next-20230720 was previously added for next-20230725
but it assumed  next-20230725 would be on 6.5.0-rc2-next but it is
actually based on  on 6.5.0-rc2-next so we remove the old symlinks
and just add the new baseline.

In order to try to make this an educational experience, I'll try
to document the process here. I simply used 'make fstests-baseline'
and then manually checked with

./scripts/workflows/fstests/fstests_watchdog.py

Until everything was done or we had a hung task. Since no expunges
were previously detected it mean that the fstests tests ran without
any expunges at all. The only ansible task which hung was the task
for xfs_rtdev guest and it was on generic/562:

./scripts/workflows/fstests/fstests_watchdog.py hosts baseline
                           Hostname           Test-name        Completion %          runtime(s)     last-runtime(s)   Stall-status                        Kernel
                     sparse-xfs-crc                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                 sparse-xfs-reflink                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
        sparse-xfs-reflink-normapbt                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
            sparse-xfs-reflink-1024                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
              sparse-xfs-reflink-2k                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
              sparse-xfs-reflink-4k                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                   sparse-xfs-nocrc                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
               sparse-xfs-nocrc-512                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                sparse-xfs-nocrc-1k                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                sparse-xfs-nocrc-2k                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                sparse-xfs-nocrc-4k                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                  sparse-xfs-logdev                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725
                   sparse-xfs-rtdev         generic/562                  0%                   26074               0   Hung-Stalled       6.5.0-rc3-next-20230725
                sparse-xfs-rtlogdev                None                  0%                   0                   0             OK       6.5.0-rc3-next-20230725

I just hit CTRL-C on the 'make fstests-baseline' run and
then ran 'make fstests-baseline-restults' which copied over
the new expunges directory I am committing here without any
changes.

You will notice generic/562 is not listed as part of the expunges
for rtdev. Upon inspection of the full log of that test though
it seems it may be a test bug, that it cannot handle rtdev or
perhaps the guest for rtdev may need more memory because the kernel
log shows some OOMs every now and then. I'll add generic/562
but after this commit.

This commit also adds the results, the results are placed in
workflows/fstests/results/6.5.0-rc3-next-20230725.xz after
i ran 'make fstests-baseline-results', as that is a path
which we add in .gitignore we must move it to our own personal
namespace for results so I put it under some arbitrary namespace
for myself under workflows/fstests/results/mcgrof/.

This path is arbitrary, I just use it so to make it clear that
I used qemu for this test and this is today's update, even though
it is for next-20230725:

mkdir -p workflows/fstests/results/mcgrof/libvirt-qemu/20230727/
cp workflows/fstests/results/6.5.0-rc3-next-20230725.xz workflows/fstests/results/mcgrof/libvirt-qemu/20230727/

So folks can reproduce:

cp .config workflows/fstests/results/mcgrof/libvirt-qemu/20230727/kdevops.config
git add workflows/fstests/results/mcgrof/libvirt-qemu/20230727/

If you enable kernel-ci the 'make fstests-baseline-loop' will scrape
the above fstests_watchdog.py output and if a stall is detected it
will kill the loop. It's a loop because the goal is to get up to the
test steady state goal of up to 100 test runs in a loop. We run a
test loop as some tests only fail after being run many times. We
aim to cover a test failure rate of about 1/100 as that means we
can run all tests in about a week's time frame. That seems sensible.

Note I used virtio drives instead of the default nvme drives for
qemu (CONFIG_LIBVIRT_EXTRA_STORAGE_DRIVE_VIRTIO=y), I did this as
virtio currently supports io-threads. An io-thread is a dedicated
process within qemu which let's it avoid having to contend the qemu
global lock. The nvme driver does not yet support io-threads but will
soon. Using io-threads is good as it means we have less false positives
on gust driver NVMe IO timeouts. The lack of io-threads on qemu creates
these NVMe IO timeouts.

We'll soon make virtio default for now, until we have qemu with
nvme and io-threads supported properly.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>